requirements.txt
torch
numpy
sentence-transformers
scikit-learn
python-dotenv